pyproject.toml
[project]
name = "modelo-recomendador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama==0.4.6",
    "evidently==0.4.26",
    "fastapi==0.108.0",
    "joblib==1.3.2",
    "loguru==0.7.2",
    "mlflow==2.9.2",
    "numpy==1.26.2",
    "pandas==2.1.3",
    "plotly==5.18.0",
    "pyarrow==14.0.1",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",
    "scikit-learn==1.3.2",
    "setuptools>=68.0.0",
    "streamlit==1.29.0",
    "torch==2.1.0",
    "tqdm==4.66.1",
    "uvicorn[standard]==0.25.0",
    "wheel>=0.41.0",
]
